Small Fix in Markdown formatting
Fixed a couple of headings that were not displaying correctly because `#` needs a space after it to display as a title.

Made them as headings; Don't know if that was the right structure of the document. Let me know if these sections were sub-headings and I'll change it to the desired structure.

Thanks for this awesome lib!